Add files via upload
Description: 
This project is a responsive website template built using HTML, CSS, Bootstrap 5, and JavaScript. It features a modern design with a navigation bar, a carousel slider, sections for About Me, Services, and a card-based layout for displaying services. The website is fully responsive and optimized for various screen sizes, providing an engaging user experience across devices. It includes Bootstrap components such as the navbar, carousel, accordion, and cards, along with custom styling to enhance visual appeal and functionality.